Added connection element for Oracle
Created element, system_instance, for Oracle connection to allow the
system user to connect to the root container, CDB$ROOT, instead of a
PDB. This allows statistics to be collected at the container level
instead of the PDB level.

A special value of 'none' means that the system will use the same
value from the 'instance' element of the Oracle connection to connect
to the System Under Test (SUT) database instance.

There is no check for consistency between the values for 'instance'
and 'system_instance' such as that they point to the container
instance.